Make ideal_answer optional in EvaluationTestCase in rag_answers/data_models.py (#74)
Updated EvaluationTestCase.ideal_answer to be optional (Optional[str])
so that certain metrics that do not require an ideal answer (faithfulness, relevancy) do not
break the evaluation pipeline when computed in isolation for test cases for which we do not have an ideal answer.

Adapted tests to handle both cases when ideal_answer is provided or None.