Add topic tag evaluation
This handles evaluating the topic tagger. It provides functionality to:

- capture the various topic categories for a single question
- aggregation of the results for all the questions in the dataset
- writing the data out to CSV files for further analysis

I've shown Nick this and he asked me to add the "Matched Any Topic"
classification, but other than that he's happy with it.

It should be noted that these tests won't pass if you run this commit alone,
but i've tried to make this PR easier to review and action comments on
by splitting it into multiple commits. The tests will pass if you check out
the branch.