Populate StructuredContext for rag_answers
Previously the code for populating retrieved_context didn't actually do
anything because the generate code wasn't configured to export source
data. This commit changes this to make use of data exported by the rake
task.

I've updated the code to use a sources field and to fail if it does not
exist. For each source this will fail if they lack a chunk field. The
reason for this change is to make it harder for this to have code that
isn't actually compatible.

As part of this I've fixed an issue where description was flagged as a
required field even though it is nullable in the DB.