Pass structured_context in LLMTestCase additional_metadata
We want to be able to pass the structured context in the desired format
for the ContextRelevancyMetric to evaluate against. We've decided on
passing it in the additional_metadata field of LLMTestCase, under the
key 'structured_context'.

We can then do whatever we need within each metric to munge the data
into the shape we want.

One thing of note, is that additional_metadata is not exposed via
LLMTestCaseParams.<CONSTANT> so i've had to add a custom validation step
in the metric to ensure that the required data is present.

I've updated retrieved_context to retrieved_contexts to reflect that
it's a list of chunks/contexts.

I've had to move the StructuredContext data model to a shared_data_models.py
file so that both data_models.py and context_relevancy.py can import it
without causing a circular import.