Refactor frozen time approach for tests
There was a lot of duplication in this approach to applying frozen time
across the similar files in this test. This moves the shared
functionality to the conftest.py file so it can be used as a single
fixture in multiple tests.

This has a disadvantage that it's no longer automatically applied to all
tests in a file, but as we only use it when we need the
mock_output_directory this doesn't seem to produce a problem.