Log each deepeval run separately to its own file
At the moment, we do n runs of deepeval evaluation, but only log the
last run to a file called deepeval_test_run.json. This means we lose the
results of the previous runs.

This moves the writing of the deepeval_test_run.json file into the
run_deepeval_evaluation function. After the evaluation is run, we get
the test run from the global_test_run_manager, and write it to the
output_dir with a filename that includes the run number.

So the first run will be logged to deepeval_test_run_1.json, the second
to deepeval_test_run_2.json, and so on.

I haven't bothered to do something like:

if n_runs > 2 :
  filename = deepeval_test_run.json
else:
  filename = deepeval_test_run_{i + 1}.json

If there's only 1 run i don't see the harm of having the _1 suffix.
Happy to change it if people disagree though.