Add optional id field to GenerateInput data model
We want to add an optional id field to the GenerateInput data model to
allow users to specify their own identifiers for input instances.
If no id is provided, a UUID will be generated by default.

This is required because sometimes we test the same question with different
contexts, and we need to be able to distinguish between these inputs in the
test results.