Remove id collisions in RAG answer evaluation
We want a way to remove collisions when multiple inputs have the same id.
This change adds an ensure_model_id_is_unique function that is called
in the generate_and_write_dataset function after GenerateInputs are created.

This function appends a suffix to any duplicate ids to ensure uniqueness.
So

if three inputs have the id "input-1", the second will be changed to
"input-1-duplicate-1" and the third to "input-1-duplicate-2".

This does mean if 2 generated UUIDs collide which is extremely likely
but possible, the second will be renamed, to uuid-duplicate-1.

I did think about doing something like:

```
for id, grouped_inputs in inputs_by_id.items():
    if len(grouped_inputs) > 1:
        for index, input in enumerate(grouped_inputs[1:], start=1):
            try:
                uuid.UUID(id)
                input.id = str(uuid.uuid4())
            except ValueError:
                input.id = f"{input.id}-duplicate-{index}"
```

inside a loop. But i'm just not convinced it's worth the additional complexity for
something that unlikely and with no real downside if it were to occur.