Remove id collisions in RAG answer evaluation
We want a way to remove collisions when multiple inputs have the same id.
This change adds an ensure_model_id_is_unique function that is called
in the generate_and_write_dataset function after GenerateInputs are created.

This adds functionality to check for id collisions when inputs are
generated. If one is found then an error is raised with the duplicate ids
so they can be added by the user.