Rename rag_answers EvaluationResult#name to id
We're going to add an id field to the GenerateInput data model to allow
users to specify unique identifiers for each generated RAG answer in the
input data.

Since we're going to be using id to uniquely identify RAG answers, it makes
sense to update the EvaluationResult data model to use id instead of name
so the output to the file is consistent with the input.