Extract ensure_model_ids_are_unique and use in generate and evaluate
There's a chance that with the current implementation, the reuse of
exisiting input files will cause collisions between EvaluationTestCase ids.

This is because they won't have been assigned an id via the generate
step.

This extracts the logic to ensure unique ids into its own function
and uses it in both generate and evaluate steps.

I'm not overly convinced we need to this, but it's mentioned in the
card to i've implemented it.

The reason i don't think we need to is because if the there aren't ids
already associated with the input, then these ids feel a bit useless anyway
so it doesn't really matter if they collide.