Add monkey patch to allow use of OpenAI OSS models in RagAnswer pipeline
After some investigation, it was found that the OpenAI OSS models on Bedrock
return multiple message contents in their responses. The first contains
a reasoning message and the second contains the actual answer.

Looking at the ContentBlock object used in the Converse API
https://docs.aws.amazon.com/bedrock/latest/APIReference/API_runtime_ContentBlock.html
we can see that it supports many content items types:

- reasoningContent
- text
- toolResult

And quite a few others.

However, looking at the DeepEval AmazonBedrockModel implementation, we see
that it assumes that the first content item is always text content:
https://github.com/confident-ai/deepeval/blob/8291dca20ac893c018012c77d47f3de477a34788/deepeval/models/llms/amazon_bedrock_model.py#L87

Due to this assumption, when the OpenAI OSS models return a reasoningContent
the DeepEval code tries to access the 'text' key on it, which does not exist,
resulting in a KeyError.

This commit adds a monkey patch to the rag_answer evaluation pipeline for the
Amazon Bedrock OpenAI OSS models. The patch modifies the 'a_measure' method
of the AmazonBedrockModel class to look for the first content item with a
'text' key and returna that as the result.

No other changes have been made to the functionality within the a_measure method.

I've tested running the pipleine using both models:

- openai.gpt-oss-20b-1:0
- openai.gpt-oss-120b-1:0

and they both seem to work as expected.

I've opened up a PR on the DeepEval repo to propose this change upstream.
See the PR description for more details.

If they respond positively and choose to merge the change, or address the
issue in a different way, we can remove this monkey patch in future.