Rename factual correctness and completeness to precision and recall
This commit renames the factual 'correctness' and 'completeness' metrics
to factual 'precision' and 'recall'. This has been agreed upon to avoid
confusion with judgements human evaluators make during manual evaluation
of judging an answer to be 'complete' or 'accurate/correct'. Using the
same terminology for both auto-eval and human judgements suggests they
measure the same thing / can be used interchangeably but they
cannot. Renaming it to 'precision' and 'recall' correctly reduces the
metrics to the statistical classification / information retrieval
definitions that those metrics implement and avoids a more conceptual
interpretation.