Loosen coherence threshold and drop 'idk' test
A threshold of 1.0 was unnecessarily strict and difficult to ensure in
practice. This caused problems with 'real_provider' tests, as it made
tests that expect a pass from the metric unreliable. In practice a
coherent answer is often scored as 4/5 by the model. Lowering the
threshold to 0.75 allows a score of 4/5 to pass.

This commit also drops the 'real_provider' test for the "I don't know
scenario". This test was also flakey, and was essentially testing an
edge case. 'real_provider' tests should be focused on the simple and
obvious test cases, not edge cases.