Verify AWS credentials are for the correct role
Chae helpfully pointed out on review that this script had become a bit
misleading when taking an AWS role as it can suggest it's assuming that
role and then say it exists when existing creds are for a different
role.

To resolve this I've added a conditional to check whether the expected
role is signed in. To do this check was a bit tricky as the role ARN
that is desired is not the same as the role one assumes. For example a
desired role has a structure of "arn:aws:iam::{account_number}:role/{role_name}"
whereas the one you assume has a structure of
"arn:aws:sts::{account_number}:assumed-role/{role_number}/{id}".

To deal with this I get the role name and account number from the `gds
aws "$role" --details` command and then do a pattern match on the
assumed role name.

My bash skills are not remotely high enough to write this script so
there is a lot of ChatGPT help here.