Pin deepeval to 3.7.5 in pyproject.toml
This is to fix dependencies trying to downgrade DeepEval as seen in this
click dependabot update [1]. We don't want to downgrade DeepEval as it's
a pretty fundamental dependency.

I'm not too sure of the conventions in Python around changing
dependencies in the pyproject.toml file. It'd be easier if we only set
the versions in pyproject.toml of a few key dependencies (like
DeepEval).

[1]: https://github.com/alphagov/govuk-chat-evaluation/pull/176